+++ maven-osgi-package-plugin/pom.xml
@@ -43,15 +43,6 @@
 
 	<dependencies>
 
-        <!-- support classes with business logic -->
-
-        <dependency>
-            <groupId>at.bestsolution</groupId>
-            <artifactId>maven-osgi-package-support</artifactId>
-            <version>0.0.1-SNAPSHOT</version>
-        </dependency>
-        <!-- =================================== -->
-
 		<dependency>
 			<groupId>org.apache.maven</groupId>
 			<artifactId>maven-core</artifactId>
