--- maven-osgi-package-support/src/main/java/at/bestsolution/maven/osgi/pack/P2RepositoryPackager.java
@@ -1,11 +1,0 @@
-package at.bestsolution.maven.osgi.pack;
-
-import org.apache.maven.project.MavenProject;
-
-/**
- *
- */
-public interface P2RepositoryPackager {
-
-    public void execute(GenerationParameters parameters);
-}
